Fix structure types
