+++ db/l2db/views.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"time"
 
-	ethCommon "github.com/ethereum/go-ethereum/common"
 	"github.com/hermeznetwork/hermez-node/common"
 	"github.com/hermeznetwork/hermez-node/common/apitypes"
 	"github.com/iden3/go-iden3-crypto/babyjub"
@@ -44,7 +43,7 @@ type PoolTxAPIView struct {
 	TokenID              common.TokenID        `meddler:"token_id"`
 	TokenItemID          uint64                `meddler:"token_item_id"`
 	TokenEthBlockNum     int64                 `meddler:"eth_block_num"`
-	TokenEthAddr         ethCommon.Address     `meddler:"eth_addr"`
+	TokenEthAddr         apitypes.HezEthAddr   `meddler:"eth_addr"`
 	TokenName            string                `meddler:"name"`
 	TokenSymbol          string                `meddler:"symbol"`
 	TokenDecimals        uint64                `meddler:"decimals"`
@@ -55,48 +54,21 @@ type PoolTxAPIView struct {
 // MarshalJSON is used to neast some of the fields of PoolTxAPIView
 // without the need of auxiliar structs
 func (tx PoolTxAPIView) MarshalJSON() ([]byte, error) {
-	type jsonToken struct {
-		TokenID          common.TokenID    `json:"id"`
-		TokenItemID      uint64            `json:"itemId"`
-		TokenEthBlockNum int64             `json:"ethereumBlockNum"`
-		TokenEthAddr     ethCommon.Address `json:"ethereumAddress"`
-		TokenName        string            `json:"name"`
-		TokenSymbol      string            `json:"symbol"`
-		TokenDecimals    uint64            `json:"decimals"`
-		TokenUSD         *float64          `json:"USD"`
-		TokenUSDUpdate   *time.Time        `json:"fiatUpdate"`
-	}
-	type jsonFormat struct {
-		ItemID               uint64                `json:"itemId"`
-		TxID                 common.TxID           `json:"id"`
-		Type                 common.TxType         `json:"type"`
-		FromIdx              apitypes.HezIdx       `json:"fromAccountIndex"`
-		EffectiveFromEthAddr *apitypes.HezEthAddr  `json:"fromHezEthereumAddress"`
-		EffectiveFromBJJ     *apitypes.HezBJJ      `json:"fromBJJ"`
-		ToIdx                *apitypes.HezIdx      `json:"toAccountIndex"`
-		EffectiveToEthAddr   *apitypes.HezEthAddr  `json:"toHezEthereumAddress"`
-		EffectiveToBJJ       *apitypes.HezBJJ      `json:"toBJJ"`
-		Amount               apitypes.BigIntStr    `json:"amount"`
-		Fee                  common.FeeSelector    `json:"fee"`
-		Nonce                common.Nonce          `json:"nonce"`
-		State                common.PoolL2TxState  `json:"state"`
-		MaxNumBatch          uint32                `json:"maxNumBatch"`
-		Info                 *string               `json:"info"`
-		ErrorCode            *int                  `json:"errorCode"`
-		ErrorType            *string               `json:"errorType"`
-		Signature            babyjub.SignatureComp `json:"signature"`
-		Timestamp            time.Time             `json:"timestamp"`
-		RqFromIdx            *apitypes.HezIdx      `json:"requestFromAccountIndex"`
-		RqToIdx              *apitypes.HezIdx      `json:"requestToAccountIndex"`
-		RqToEthAddr          *apitypes.HezEthAddr  `json:"requestToHezEthereumAddress"`
-		RqToBJJ              *apitypes.HezBJJ      `json:"requestToBJJ"`
-		RqTokenID            *common.TokenID       `json:"requestTokenId"`
-		RqAmount             *apitypes.BigIntStr   `json:"requestAmount"`
-		RqFee                *common.FeeSelector   `json:"requestFee"`
-		RqNonce              *common.Nonce         `json:"requestNonce"`
-		Token                jsonToken             `json:"token"`
-	}
-	toMarshal := jsonFormat{
+	toMarshal := tx.ToAPI()
+	return json.Marshal(toMarshal)
+}
+
+// AccountCreationAuthAPI represents an account creation auth in the expected format by the API
+type AccountCreationAuthAPI struct {
+	EthAddr   apitypes.HezEthAddr   `json:"hezEthereumAddress" meddler:"eth_addr" `
+	BJJ       apitypes.HezBJJ       `json:"bjj"                meddler:"bjj" `
+	Signature apitypes.EthSignature `json:"signature"          meddler:"signature" `
+	Timestamp time.Time             `json:"timestamp"          meddler:"timestamp,utctime"`
+}
+
+// ToAPI converts a PoolTxAPIView into PoolL2TxAPI
+func (tx *PoolTxAPIView) ToAPI() apitypes.PoolL2TxAPI {
+	pooll2apilocal := apitypes.PoolL2TxAPI{
 		ItemID:               tx.ItemID,
 		TxID:                 tx.TxID,
 		Type:                 tx.Type,
@@ -124,25 +96,16 @@ func (tx PoolTxAPIView) MarshalJSON() ([]byte, error) {
 		RqAmount:             tx.RqAmount,
 		RqFee:                tx.RqFee,
 		RqNonce:              tx.RqNonce,
-		Token: jsonToken{
-			TokenID:          tx.TokenID,
-			TokenItemID:      tx.TokenItemID,
-			TokenEthBlockNum: tx.TokenEthBlockNum,
-			TokenEthAddr:     tx.TokenEthAddr,
-			TokenName:        tx.TokenName,
-			TokenSymbol:      tx.TokenSymbol,
-			TokenDecimals:    tx.TokenDecimals,
-			TokenUSD:         tx.TokenUSD,
-			TokenUSDUpdate:   tx.TokenUSDUpdate,
-		},
 	}
-	return json.Marshal(toMarshal)
-}
+	pooll2apilocal.Token.TokenID = tx.TokenID
+	pooll2apilocal.Token.TokenItemID = tx.TokenItemID
+	pooll2apilocal.Token.TokenEthBlockNum = tx.TokenEthBlockNum
+	pooll2apilocal.Token.TokenEthAddr = tx.TokenEthAddr
+	pooll2apilocal.Token.TokenName = tx.TokenName
+	pooll2apilocal.Token.TokenSymbol = tx.TokenSymbol
+	pooll2apilocal.Token.TokenDecimals = tx.TokenDecimals
+	pooll2apilocal.Token.TokenUSD = tx.TokenUSD
+	pooll2apilocal.Token.TokenUSDUpdate = tx.TokenUSDUpdate
 
-// AccountCreationAuthAPI represents an account creation auth in the expected format by the API
-type AccountCreationAuthAPI struct {
-	EthAddr   apitypes.HezEthAddr   `json:"hezEthereumAddress" meddler:"eth_addr" `
-	BJJ       apitypes.HezBJJ       `json:"bjj"                meddler:"bjj" `
-	Signature apitypes.EthSignature `json:"signature"          meddler:"signature" `
-	Timestamp time.Time             `json:"timestamp"          meddler:"timestamp,utctime"`
+	return pooll2apilocal
 }
